% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxZarr-class.R
\docType{class}
\name{TENxZarr-class}
\alias{TENxZarr-class}
\alias{.TENxZarr}
\title{A minimal class to represent Zarr files}
\value{
A \code{TENxZarr} class object
}
\description{
A minimal class to represent Zarr files
}
