% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxZarr-class.R
\name{TENxZarr}
\alias{TENxZarr}
\alias{decompress,TENxZarr-method}
\alias{import,TENxZarr,ANY,ANY-method}
\title{Import 10X data from the Zarr format}
\usage{
TENxZarr(resource)

\S4method{decompress}{TENxZarr}(manager, con, ...)

\S4method{import}{TENxZarr,ANY,ANY}(con, format, text, ...)
}
\arguments{
\item{resource}{\code{character(1)} The path to the Zarr file. Can be zipped.}
}
\value{
A \code{TENxZarr} class object
}
\description{
Import 10X data from the Zarr format
}
\examples{
zarrzip <- paste0(
    "~/data/Xenium_Prime_MultiCellSeg_Mouse_Ileum_tiny_outs/",
    "cell_feature_matrix.zarr.zip"
)
TENxZarr(zarrzip) |>
    import()

}
