% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XeniumIO-package.R
\docType{package}
\name{XeniumIO-package}
\alias{XeniumIO}
\alias{XeniumIO-package}
\title{XeniumIO: Import and represent Xenium data from the 10X Xenium Analyzer}
\description{
The package allows users to readily import spatial data obtained from the 10X Xenium Analyzer pipeline. Supported formats include 'parquet', 'h5', and 'mtx' files. The package mainly represents data as SpatialExperiment objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/waldronlab/XeniumIO}
  \item Report bugs at \url{https://github.com/waldronlab/XeniumIO/issues}
}

}
\author{
\strong{Maintainer}: Marcel Ramos \email{marcel.ramos@sph.cuny.edu} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})

Other contributors:
\itemize{
  \item Dario Righelli [contributor]
  \item Estella Dong [contributor]
}

}
