% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxXenium-class.R
\docType{class}
\name{TENxXenium-class}
\alias{TENxXenium-class}
\alias{.TENxXenium}
\alias{TENxXenium}
\title{A class to represent Xenium output data}
\usage{
TENxXenium(
  resources,
  xeniumOut,
  sample_id = "sample01",
  format = c("mtx", "h5"),
  spatialCoordsNames = c("x_centroid", "y_centroid"),
  ...
)
}
\arguments{
\item{resources}{A \link{TENxFileList} object or a file path to the tarball
containing the matrix / assay data resources.}

\item{xeniumOut}{\code{character(1)} The path to the Xenium output directory.}

\item{sample_id}{\code{character(1)} A single string specifying the sample ID.}

\item{format}{The format of the output. If missing and \code{con} is a
file name, the format is derived from the file extension. This argument is
unnecessary when \code{con} is a derivative of \linkS4class{BiocFile}.}

\item{spatialCoordsNames}{\code{character()} A vector of strings specifying the
names of the columns in the spatial data containing the spatial coordinates.}

\item{...}{In the constructor, additional arguments passed to
\code{\link[=TENxFileList]{TENxFileList()}}; otherwise, not used.}
}
\value{
A \link{SpatialExperiment} object
}
\description{
This class is a composed class of \link{TENxFileList} which can
contain a list of \link{TENxFile} objects for the cell-feature matrix. It is
meant to handle a single Xenium sample from 10X Genomics.
}
\section{Slots}{

\describe{
\item{\code{resources}}{A \link{TENxFileList} or \link{TENxH5} object containing the cell
feature matrix.}

\item{\code{coordNames}}{\code{character()} A vector specifying the names
of the columns in the spatial data containing the spatial coordinates.}

\item{\code{sampleId}}{\code{character(1)} A scalar specifying the sample identifier.}
}}

\seealso{
\url{https://www.10xgenomics.com/support/software/xenium-onboard-analysis/latest/analysis/xoa-output-understanding-outputs}
}
